move cache into a separate file
